# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FarmaSystem - API de Gestão de Farmácia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock:
    get:
      operationId: listStock
      tags: [stock]
      summary: Listar lotes de estoque
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockLot"
    post:
      operationId: createStockEntry
      tags: [stock]
      summary: Entrada de mercadoria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockEntryBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLot"
  /stock/movements:
    get:
      operationId: listStockMovements
      tags: [stock]
      summary: Histórico de movimentações
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
  /stock/adjust:
    post:
      operationId: adjustStock
      tags: [stock]
      summary: Ajuste manual de estoque
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"

  # ─── Sales ────────────────────────────────────────────────────
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: Listar fornecedores
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Criar fornecedor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Obter fornecedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Atualizar fornecedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Excluir fornecedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Excluído

  # ─── Prescriptions ────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Listar usuários
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Criar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Atualizar usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Excluir usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Excluído
components:
  schemas:
...
    StockLot:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        lotNumber:
          type: string
        quantity:
          type: integer
        expirationDate:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        receivedAt:
          type: string
        costPrice:
          type: number
      required:
        - id
        - productId
        - productName
        - lotNumber
        - quantity
        - expirationDate
        - receivedAt
        - costPrice
    CreateStockEntryBody:
      type: object
      properties:
        productId:
          type: integer
        lotNumber:
          type: string
        quantity:
          type: integer
        expirationDate:
          type: string
        supplierId:
          type: ["integer", "null"]
        costPrice:
          type: number
      required:
        - productId
        - lotNumber
        - quantity
        - expirationDate
        - costPrice
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        type:
          type: string
          enum: [entrada, saida, ajuste, transferencia]
        quantity:
          type: integer
        reason:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - type
        - quantity
        - userId
        - userName
        - createdAt
    StockAdjustBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        reason:
          type: string
      required:
        - productId
        - quantity
        - reason
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cnpj:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        contact:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        cnpj:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        contact:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required:
        - name
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, gerente, atendente, farmaceutico]
        active:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - active
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, gerente, atendente, farmaceutico]
        active:
          type: boolean
      required:
        - name
        - email
        - role